include: Get the include files out of the way of the main code.
This way, we don't see a polluted source directory. We adapt the inclusion
of the header files accordingly. The build system is flaky at the moment,
but that will be dealt with in the near future (hopefully!).

One future next step is to move the source code under a `src` directory.