Redesign of main function
Now it takes arguments from command line specifying size of array and
matrix to be used. This data is generated at execution time by gen_data
function in auxiliary.{c,h}.

There is also some basic error treatment argument that should be
improved, specially the logic expression in the first if statement. It
seems to produce desired results, but produces a segmentation fault when
less then 3 arguments are passed to the program. Maybe logical operators
are not associative in C and it has to evaluate every logic expression
inside if statement. If so it tries to access invalid position in
argv[]. Suggestions?